scalable e2e tests
