+++ e2e/entrypoint.sh
@@ -47,7 +47,7 @@ echo
 
 if [[ "$os" == "linux" ]]
 then
-    xvfb-run -a -e /dev/stdout -s "-screen 0 1280x960x24" yarn playwright test api.spec.ts
+    xvfb-run -a -e /dev/stdout -s "-screen 0 1280x960x24" yarn playwright test .*.spec.ts
 else
-    yarn playwright test api.spec.ts
+    yarn playwright test .*.spec.ts
 fi
